+++ cpskin/core/setuphandlers.py
@@ -1,9 +1,11 @@
 # -*- coding: utf-8 -*-
+from collective.dexteritytextindexer.behavior import IDexterityTextIndexer
 from collective.taxonomy.exportimport import TaxonomyImportExportAdapter
 from collective.taxonomy.factory import registerTaxonomy
 from cpskin.core.behaviors.directory import ICpskinDirectoryViewSettings
 from cpskin.core.behaviors.eventview import ICpskinEventViewSettings
 from cpskin.core.behaviors.indexview import ICpskinIndexViewSettings
+from cpskin.core.behaviors.metadata import IAdditionalSearchableText
 from cpskin.core.behaviors.organization import IOrganizationImages
 from cpskin.core.utils import add_behavior
 from cpskin.core.utils import set_exclude_from_nav
@@ -54,6 +56,8 @@ def installCore(context):
     add_behavior('Collection', ICpskinIndexViewSettings.__identifier__)
     add_behavior('Event', ICpskinEventViewSettings.__identifier__)
     add_behavior('organization', IOrganizationImages.__identifier__)
+    add_behavior('organization', IAdditionalSearchableText.__identifier__)
+    add_behavior('organization', IDexterityTextIndexer.__identifier__)
     add_behavior('directory', ICpskinDirectoryViewSettings.__identifier__)
 
     # Create footer static page